Update README.md with instruction for R dependencies
Addresses gap in not having an R dependency script. The R dependencies are well-formatted in the HM however they still need to be passed in as a list into the manual download_deps.R script (TODO).

Also TODO update the formatting of the python HM as the proper structure is now defined in the R-generation script.